+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Function for opening windows/popups
+
 ## [v2.6.1] - 2020-12-18
 
 ### Added