+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Function for opening windows/popups
+- Function to get files from URLs (using `Content-Type` header)
+- Functions for `fetch`ing `<link>`s (useful for preloads & manifests)
+- Add function for async submitting of forms
 
 ## [v2.6.1] - 2020-12-18